# swagger.yaml
openapi: 3.0.0
info:
  title: invaPlay Backend API
  version: 1.0.0
  description: API documentation for the invaPlay Backend application, managing users, students, schools, classes, and grades.
servers:
  - url: http://localhost:5000 # Assuming the API is served from the root
  - url: https://api.innovplay.online/
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /test/{test_id}:
    get:
      summary: Get test details by ID
      parameters:
        - name: test_id
          in: path
          required: true
          schema:
            type: string
          description: ID of the test
      responses:
        '200':
          description: Test details
          content:
            application/json:
              schema:
                $ref: '#/schemas/Test'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
        '404':
          $ref: '#/responses/NotFoundError'
        '500':
          $ref: '#/responses/InternalServerError'
    put:
      summary: Update test by ID
      parameters:
        - name: test_id
          in: path
          required: true
          schema:
            type: string
          description: ID of the test to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                type:
                  type: string
                model:
                  type: string
                course:
                  type: string
                description:
                  type: string
                subject:
                  type: string
                grade:
                  type: string
                grade_id:
                  type: string
                intructions:
                  type: string
                max_score:
                  type: number
                time_limit:
                  type: string
                  format: date-time
                municipalities:
                  type: array
                schools:
                  type: array
                subjects:
                  type: array
                subjects_info:
                  type: array
      responses:
        '200':
          description: Test updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/responses/BadRequestError'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
        '404':
          $ref: '#/responses/NotFoundError'
        '500':
          $ref: '#/responses/InternalServerError'
    delete:
      summary: Delete a test by ID
      parameters:
        - name: test_id
          in: path
          required: true
          schema:
            type: string
          description: ID of the test to delete
      responses:
        '200':
          description: Test deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
        '404':
          $ref: '#/responses/NotFoundError'
        '500':
          $ref: '#/responses/InternalServerError'
...
  /student-answers/student/tests/status:
    get:
      summary: Get all student tests status
      responses:
        '200':
          description: All student tests status
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    test_id:
                      type: string
                    test_title:
                      type: string
                    status:
                      type: string
                    can_start:
                      type: boolean
                    started_at:
                      type: string
                      format: date-time
                      nullable: true
                    submitted_at:
                      type: string
                      format: date-time
                      nullable: true
                    score:
                      type: number
                      nullable: true
                    grade:
                      type: number
                      nullable: true
        '401':
          $ref: '#/responses/UnauthorizedError'
        '500':
          $ref: '#/responses/InternalServerError'
...
responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Mensagem explicando o motivo da falha de autenticação.
                example: Token inválido ou expirado.
    ForbiddenError:
      description: User does not have the necessary permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Mensagem indicando que o usuário não tem permissão.
                example: Acesso negado para o recurso solicitado.
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Mensagem sobre o recurso não encontrado.
                example: Recurso solicitado não foi encontrado.
    BadRequestError:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Mensagem descrevendo o erro de validação.
                example: Dados inválidos enviados na requisição.
              details:
                type: object # Can be a string or a more detailed object
                description: Detalhes adicionais sobre o erro.
                example:
                  campo: email
                  problema: Formato inválido.
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Mensagem genérica de erro.
                example: Ocorreu um erro inesperado.
              details:
                type: string
                description: Descrição detalhada do erro para depuração.
                example: Consulte os logs do servidor para mais informações.